Fixed some Wtf-issues in backend

+++ areas/admin/__init__.py
@@ -1,10 +1,7 @@
-from flask import Blueprint, render_template, request
+from flask import Blueprint, render_template, request, redirect
 from flask_security import roles_accepted
-from flask import (
-    render_template,
-    request,
-    redirect,
-)
+
+from forms import DeleteNewsletterEmailForm
 from models import NewsletterEmails, db
 
 adminBluePrint = Blueprint("admin", __name__)
@@ -19,8 +16,11 @@ def admin():
 @adminBluePrint.route("/manage-newsletter", methods=["GET", "POST"])
 @roles_accepted("Admin", "Staff")
 def manage_newsletter():
+    form_delete = DeleteNewsletterEmailForm(request.form)
     return render_template(
-        "admin/manageNewsletter.html", values=NewsletterEmails.query.all()
+        "admin/manageNewsletter.html",
+        values=NewsletterEmails.query.all(),
+        form_delete=form_delete,
     )
 
 
@@ -30,14 +30,19 @@ def write_newsletter():
     return render_template("admin/writeNewsletter.html")
 
 
-@adminBluePrint.post("/delete-email")
+@adminBluePrint.route("/delete-email", methods=["POST"])
 @roles_accepted("Admin", "Staff")
 def delete_email():
-    email_id = int(request.form["email_id"])
-    email_to_delete = NewsletterEmails.query.get(email_id)
-    if email_to_delete:
-        db.session.delete(email_to_delete)
-        db.session.commit()
-        return redirect("/admin/manage-newsletter")
+    if request.method == "POST":
+        form_delete = DeleteNewsletterEmailForm(request.form)
+        if form_delete.validate():  # Validerar formuläret inklusive email_id
+            email_id = form_delete.email_id.data
+            email_to_delete = NewsletterEmails.query.get(email_id)
+            if email_to_delete:
+                db.session.delete(email_to_delete)
+                db.session.commit()
+                return redirect("/admin/manage-newsletter")
+        else:
+            return "Email not found", 404
     else:
-        return "Email not found", 404
+        return "Invalid form data", 400
